decimal start - save logo complete

+++ app/src/main/java/com/skywavestudios/prefactor/Product.java
@@ -7,6 +7,6 @@ package com.skywavestudios.prefactor;
 public class Product {
     public String Name;
     public int Fee;
-    public int No;
+    public float No;
     public String Scale;
 }
